PEPPER-1402 more refactoring, changed EventUtil to EventService
